exercises for period 2

+++ Periode2/express-exercises-day1/src/routes/userApi.ts
@@ -1,58 +1,94 @@
 import express from "express";
 import userFacade from "../facades/user";
-import basicAuth from "../middlewares/basic-auth";
-import {ApiError2} from "../errors/apiError2";
-import authMiddleware from "../middlewares/basic-auth";
+import basicAuth from "../middlewares/basic-auth"
 const router = express.Router();
+import { ApiError } from "../errors/apiError"
+import authMiddleware from "../middlewares/basic-auth";
+
+
 
 router.post('/', async function (req, res, next) {
-  try{
+  try {
     let newUser = req.body;
     newUser.role = "user";  //Even if a hacker tried to "sneak" in his own role, this is what you get
     const status = await userFacade.addUser(newUser)
-    res.json({status})
-  } catch (err){
-    throw new ApiError2(err.message, 404, res)
+    res.json({ status })
+  } catch (err) {
     next(err);
   }
 })
 
-router.use(authMiddleware)
+const useAuthentication = !process.env["SKIP_AUTHENTICATION"];
+if (useAuthentication) {
+  router.use(authMiddleware)
+}
 
-router.get('/:userName', async function (req, res, next) {
+
+router.get('/:userName', async function (req: any, res, next) {
   try {
-  const user_Name = req.params.userName;
-  const user = await userFacade.getUser(user_Name);
-  const { name, userName } = user;
-  const userDTO = { name, userName }
-  res.json(userDTO);
-  } catch(err){
-    throw new ApiError2(err.message, 404, res)
+    const role = req.role;
+    if (useAuthentication && role != "admin") {
+      throw new ApiError("Not Authorized", 403)
+    }
+    const user_Name = req.params.userName;
+    const user = await userFacade.getUser(user_Name);
+    const { name, userName } = user;
+    const userDTO = { name, userName }
+    res.json(userDTO);
+  } catch (err) {
     next(err)
   }
 });
 
-router.get('/', async function (req, res, next) {
+
+router.get('/user/me', async function (req: any, res, next) {
+  if (!useAuthentication) {
+    throw new ApiError("Method cannot be used without authentication")
+  }
   try {
+    const user_Name = req.userName;
+    const user = await userFacade.getUser(user_Name);
+    const { name, userName } = user;
+    const userDTO = { name, userName }
+    res.json(userDTO);
+  } catch (err) {
+    next(err)
+  }
+});
+
+
+
+router.get('/', async function (req: any, res, next) {
+  try {
+
+    const role = req.role;
+    if (useAuthentication && role != "admin") {
+      throw new ApiError("Not Authorized", 403)
+    }
+
     const users = await userFacade.getAllUsers();
-    const usersDTO = users.map((user) => { 
+    const usersDTO = users.map((user) => {
       const { name, userName } = user;
-      return {name,userName}
+      return { name, userName }
     })
     res.json(usersDTO);
   } catch (err) {
-    throw new ApiError2(err.message, 404, res)
     next(err)
   }
 });
 
-router.delete('/:userName', async function (req, res, next) {
-  try{
+router.delete('/:userName', async function (req: any, res, next) {
+  try {
+
+    const role = req.role;
+    if (useAuthentication && role != "admin") {
+      throw new ApiError("Not Authorized", 403)
+    }
+
     const user_name = req.params.userName;
     const status = await userFacade.deleteUser(user_name)
-    res.json({status})
-  } catch (err){
-    throw new ApiError2(err.message, 404, res)
+    res.json({ status })
+  } catch (err) {
     next(err);
   }
 })
